Split locations into a directory
