+++ ocelot/transformations/locations/__init__.py
@@ -1,0 +1,4 @@
+# -*- coding: utf-8 -*-
+from .topology import Topology
+from .rest_of_world import relabel_global_to_row
+from .validation import check_single_global_dataset
